Move DatabaseSearch to phpunit Integration tests
