+++ src/OpenCATS/Tests/IntegrationTests/DatabaseSearchTest.php
@@ -1,0 +1,80 @@
+<?php
+namespace OpenCATS\Tests\IntegrationTests;
+
+use \OpenCATS\Tests\IntegrationTests\DatabaseTestCase;
+use DatabaseConnection;
+use DatabaseSearch;
+
+include_once('./lib/DatabaseSearch.php');
+
+class DatabaseSearchTest extends DatabaseTestCase
+{
+    function testMakeREGEXPString()
+    {
+        //FIXME: Write me!
+    }
+
+    function testMakeBooleanSQLWhere()
+    {
+        $tests = array(
+            array(
+                'java',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\'))'
+            ),
+            array(
+                'java sql',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\') AND (field REGEXP \'[[:<:]]sql[[:>:]]\'))'
+            ),
+            array(
+                'java | sql',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\') OR (field REGEXP \'[[:<:]]sql[[:>:]]\'))'
+            ),
+            array(
+                'java,sql',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\') OR (field REGEXP \'[[:<:]]sql[[:>:]]\'))'
+            ),
+            array(
+                'java, ,,sql',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\') OR (field REGEXP \'[[:<:]]sql[[:>:]]\'))'
+            ),
+            array(
+                'java -sql',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\') AND NOT (field REGEXP \'[[:<:]]sql[[:>:]]\'))'
+            ),
+            array(
+                'java !sql',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\') AND NOT (field REGEXP \'[[:<:]]sql[[:>:]]\'))'
+            ),
+            array(
+                'java*',
+                '((field LIKE \'%java%\'))'
+            ),
+            array(
+                'java* sql*',
+                '((field LIKE \'%java%\') AND (field LIKE \'%sql%\'))'
+            ),
+            array(
+                'java (',
+                '0'
+            ),
+            array(
+                'java) (',
+                '0'
+            ),
+            array(
+                'java ()',
+                '((field REGEXP \'[[:<:]]java[[:>:]]\'))'
+            )
+        );
+
+        $db = DatabaseConnection::getInstance();
+        foreach ($tests as $test)
+        {
+            $this->assertSame(
+                DatabaseSearch::makeBooleanSQLWhere($test[0], $db, 'field'),
+                $test[1]
+                );
+        }
+    }
+}
+?>
